fix(compile): detect archive format from object bytes, not the file name
The well-known-wrapper strip-dedup path decided COFF from the archive's
`.lib` extension. It is handed perry's own intermediate, which
`strip_duplicate_objects_from_no_shared_deps` names
`_<lib>_nosharedeps.lib` on EVERY host — so macOS and Linux took the
Windows branch, `llvm-ar --format=coff` wrote a GNU-style symbol table,
and `ld` rejected the archive:

    ld: archive member '/' not a mach-o file in _libperry_ext_ws.a_nosharedeps.lib

Read the container from an extracted member's magic bytes instead
(`object_format::object_is_coff`), so `--redefine-sym` / `--format=coff`
stay Windows-only. Full OpenCode source-graph build now links on macOS:
2,954 modules, 0 JavaScript fallbacks, 235 MB executable.

Also in this commit:

- Add the missing `runtime_erased` field to five `Import` literals in
  perry-transform's tests; the crate's lib tests did not compile, which
  red-lined cargo-test and fail-fast cancelled check / e2e-scoped /
  gc-stress-build / gap-suite-build before they ran.
- Split five files back under the 2,000-line gate: error.rs,
  property_get.rs, module_decl.rs, cjs_wrap/tests.rs and strip_dedup.rs.
- Guard the new ClassRef `Object.setPrototypeOf` store with
  `is_above_handle_band` + `try_read_gc_header` before `is_valid_obj_ptr`.
  `proto` is user-supplied and lands in a GC root table the collector
  later dereferences; a bare check accepts the fetch/zlib/proxy handle
  bands on Linux. Clears the addr-class ratchet.
- Record the `test_only` verdict for TEST_COLLECT_BEFORE_GLOBAL_THIS_ALLOC
  in the GC root-holder inventory.

Verified: cargo fmt --check, cargo check --all-targets, file-size gate,
addr-class audit and gc_runtime_root_holders all clean; perry-transform
121, perry-hir 364, perry-runtime 2,825, perry bin 1,068 tests pass.
