+++ crates/perry-hir/src/lower/module_decl.rs
@@ -13,6 +13,7 @@ mod namespace;
 mod native_default_import;
 pub(super) mod native_profile_import;
 mod object_literal;
+mod static_import_bindings;
 mod typescript;
 
 // Re-export moved items so existing `crate::...` / `super::*` call paths keep
@@ -23,70 +24,9 @@ use native_default_import::{
     node_submodule_default_export_key,
 };
 use object_literal::is_direct_object_literal;
-
-/// Register ordinary source-module import bindings before statement lowering.
-///
-/// ESM imports are module-scoped and hoisted regardless of where their
-/// declarations appear in the source. The main declaration pass still emits
-/// the HIR `Import` records in source order; this pre-pass only makes the
-/// bindings visible to expressions that precede the declaration.
-pub(super) fn pre_register_static_import_bindings(
-    ctx: &mut LoweringContext,
-    ast_module: &ast::Module,
-) {
-    for item in &ast_module.body {
-        let ast::ModuleItem::ModuleDecl(ast::ModuleDecl::Import(import_decl)) = item else {
-            continue;
-        };
-        let raw_source = import_decl.src.value.as_str().unwrap_or("").to_string();
-        let source = canonicalize_native_import_source(&raw_source);
-
-        // Native imports need their module/method-specific registration, which
-        // the ordinary declaration pass performs. This pass fixes source
-        // modules, whose value bindings all share the imported-function path.
-        if is_native_module(&source)
-            || is_node_builtin_module(&source)
-            || source == "reflect-metadata"
-        {
-            continue;
-        }
-
-        for specifier in &import_decl.specifiers {
-            match specifier {
-                ast::ImportSpecifier::Named(named) => {
-                    if import_decl.type_only || named.is_type_only {
-                        continue;
-                    }
-                    let local = named.local.sym.to_string();
-                    ctx.register_imported_func(local.clone(), local);
-                }
-                ast::ImportSpecifier::Default(default) => {
-                    if import_decl.type_only {
-                        continue;
-                    }
-                    let local = default.local.sym.to_string();
-                    ctx.register_imported_func(local.clone(), local.clone());
-                    if source == "react" {
-                        ctx.react_default_import_local = Some(local);
-                    }
-                }
-                ast::ImportSpecifier::Namespace(namespace) => {
-                    if import_decl.type_only {
-                        continue;
-                    }
-                    let local = namespace.local.sym.to_string();
-                    ctx.register_imported_func(local.clone(), local.clone());
-                    ctx.namespace_import_locals.insert(local.clone());
-                    ctx.namespace_import_sources
-                        .insert(local.clone(), source.clone());
-                    if source == "react" {
-                        ctx.react_default_import_local = Some(local);
-                    }
-                }
-            }
-        }
-    }
-}
+pub(super) use static_import_bindings::{
+    import_is_runtime_erased, pre_register_static_import_bindings,
+};
 
 pub(crate) fn lower_module_decl(
     ctx: &mut LoweringContext,
@@ -192,23 +132,7 @@ pub(crate) fn lower_module_decl(
                 return Ok(());
             }
             let whole_decl_type_only = import_decl.type_only;
-            // TypeScript's per-specifier spelling
-            //
-            //   import { type Foo, type Bar } from "./types"
-            //
-            // is just as runtime-erased as `import type { Foo, Bar }`. Keep
-            // the named specifiers below so class/interface metadata can still
-            // reach consumers, but mark the declaration itself type-only when
-            // every specifier is erased. Mixed imports retain their runtime
-            // edge because at least one specifier carries a value binding.
-            let runtime_erased = !whole_decl_type_only
-                && !import_decl.specifiers.is_empty()
-                && import_decl.specifiers.iter().all(|specifier| {
-                    matches!(
-                        specifier,
-                        ast::ImportSpecifier::Named(named) if named.is_type_only
-                    )
-                });
+            let runtime_erased = import_is_runtime_erased(import_decl, whole_decl_type_only);
 
             // Parse import specifiers
             let mut specifiers = Vec::new();
