+++ crates/perry-transform/src/inline/mod.rs
@@ -1093,6 +1093,7 @@ mod tests {
             module_kind: ModuleKind::NativeCompiled,
             resolved_path: Some("/src/predicate.ts".to_string()),
             type_only: false,
+            runtime_erased: false,
             is_dynamic: false,
             is_dynamic_target: false,
             is_deferred_require: false,
@@ -1151,6 +1152,7 @@ mod tests {
             module_kind: ModuleKind::NativeCompiled,
             resolved_path: Some("/src/ops.ts".to_string()),
             type_only: false,
+            runtime_erased: false,
             is_dynamic: false,
             is_dynamic_target: false,
             is_deferred_require: false,
@@ -1166,6 +1168,7 @@ mod tests {
             module_kind: ModuleKind::NativeCompiled,
             resolved_path: Some("/src/predicate.ts".to_string()),
             type_only: false,
+            runtime_erased: false,
             is_dynamic: false,
             is_dynamic_target: false,
             is_deferred_require: false,
@@ -1334,6 +1337,7 @@ mod tests {
             module_kind: ModuleKind::NativeCompiled,
             resolved_path: Some("/src/ops.ts".to_string()),
             type_only: false,
+            runtime_erased: false,
             is_dynamic: false,
             is_dynamic_target: false,
             is_deferred_require: false,
@@ -1349,6 +1353,7 @@ mod tests {
             module_kind: ModuleKind::NativeCompiled,
             resolved_path: Some("/src/predicate.ts".to_string()),
             type_only: true,
+            runtime_erased: false,
             is_dynamic: false,
             is_dynamic_target: false,
             is_deferred_require: false,
