+++ crates/perry/src/commands/compile/cjs_wrap/tests.rs
@@ -14,6 +14,7 @@ use super::wrap::{wrap_commonjs, wrap_commonjs_for_target, wrap_commonjs_with_bo
 use std::fs;
 use std::path::PathBuf;
 
+mod hoist_scanner;
 mod source_graph;
 
 // #5247: the wrapped output must report where the ORIGINAL body begins, and
@@ -1983,32 +1984,5 @@ fn cjs_wrap_builtin_require_not_hoisted_as_static_import() {
     assert!(
         !wrapped.contains("return _req_0;"),
         "the built-in require case must not reference the dropped import local; got:\n{wrapped}"
-    );
-}
-
-#[test]
-fn regex_quote_before_local_superclass_keeps_class_in_cjs_iife() {
-    // @smithy/core's serde CJS emit contains this sequence. The quote inside
-    // the regex is not a string delimiter; treating it as one desynchronized
-    // the top-level-binding scanner, hid `ReadableStreamRef`, and hoisted only
-    // `ChecksumStream` ahead of the CommonJS IIFE.
-    let src = r#"const splitHeader = (value) => {
-    return value.replace(/\\"/g, '"');
-};
-const ReadableStreamRef = typeof ReadableStream === "function"
-    ? ReadableStream
-    : function () {};
-class ChecksumStream extends ReadableStreamRef {}
-module.exports = { ChecksumStream };
-"#;
-
-    let (blocks, hoisted_names, rest) = extract_top_level_class_decls(src);
-    assert!(
-        !hoisted_names.iter().any(|name| name == "ChecksumStream"),
-        "a class depending on a CJS-local superclass must not hoist; hoisted block:\n{blocks}"
-    );
-    assert!(
-        rest.contains("class ChecksumStream extends ReadableStreamRef"),
-        "the class declaration must remain at its source position inside the factory"
     );
 }
